Using the Arraylist the asteroids, spaceship, and bullets are added
Collision code is added to the asteroids so when the players get hit by the asteroids they die

Shooting mechanism is added to the player where the player can now use the shooting mechanics to destroy the asteroids

Earth and Moving stars on the background is added up to the display

Death mechanism and Respawn mechanism will be added in the milestone 4 next week

Ths major changes for this milestone is addition of arraylist and new assets added such as background drawings and asteroids, and the addition of collision codes for the spaceship and the asteroids. The death mechanics is not added yet, but in the next milestone the rest of the missing inventory skills and mechanics will be added for the final touch.