I am quite stuck in this milestone, not sure how to add up the inventory skills in the code.
I don't think I can finish up in this milestone submission, but by tomorrow before the final due date I can figure out how to complete the code

The collision code is a bit adjusted in size because the collision size is too big compared to the display of asteroids.

Planning to add up another type of asteroid, so it's more entertaining to the visual. Another shape of asteroids would improve the visuals due to the diversity of asteroids

changed up the colors of the stars and the player